Verificações CRUD BL API e componente React Sales
O que foi feito desde o último commit:
- Adicionadas verificações de erros e de integridade de dados às funções do CRUD das Sales na BL API;
- Adicionado componente do React para fazer o CRUD no Front-End.

Ricardo Gonçalves, 23/01/2024 - 17:02

+++ src/bl-api/src/sales/sales.service.ts
@@ -1,5 +1,6 @@
 import {Injectable} from '@nestjs/common';
 import {PrismaClient, Sales} from '@prisma/client';
+import {isValidDate} from "rxjs/internal/util/isDate";
 
 export type SaleData = Partial<Sales> & { id: string };
 
@@ -15,6 +16,18 @@ export class SalesService {
         if (!saleData.id) {
             throw new Error('id is required');
         }
+        if (!isValidDate(saleData.cal_month_num)) {
+            throw new Error('valid month number is required');
+        }
+        if (!isValidDate(saleData.calendar_year)) {
+            throw new Error('valid calendar year is required');
+        }
+        const existingSale = await this.prisma.sales.findUnique({
+            where: {id: saleData.id}
+        });
+        if (existingSale) {
+            throw new Error('a sale with this id already exists');
+        }
         saleData.id = saleData.id.toString();
 
         return this.prisma.sales.create({
@@ -26,6 +39,18 @@ export class SalesService {
         if (!saleData.id) {
             throw new Error('id is required');
         }
+        if (!isValidDate(saleData.cal_month_num)) {
+            throw new Error('valid month number is required');
+        }
+        if (!isValidDate(saleData.calendar_year)) {
+            throw new Error('valid calendar year is required');
+        }
+        const existingSale = await this.prisma.sales.findUnique({
+            where: {id: saleData.id}
+        });
+        if (!existingSale) {
+            throw new Error('a sale with this id does not exist');
+        }
         saleData.id = saleData.id.toString();
 
         return this.prisma.sales.update({
@@ -34,6 +59,15 @@ export class SalesService {
         });
     }
     async deleteSale(id: string): Promise<Sales> {
+        if (!id) {
+            throw new Error('id is required');
+        }
+        const existingSale = await this.prisma.sales.findUnique({
+           where: {id: id}
+        });
+        if (!existingSale) {
+            throw new Error('a sale with this id does not exist');
+        }
         return this.prisma.sales.delete({
             where: {id: id}
         });
